move things out of menu_tags.py that don't belong there
