+++ corehq/apps/hqwebapp/templatetags/menu_tags.py
@@ -1,16 +1,10 @@
-import logging
 from django import template
-from django.conf import settings
-from django.core.urlresolvers import reverse
 from django.template.loader import render_to_string
-from django.templatetags.i18n import language_name
-from django.utils.html import format_html
 from django.utils.safestring import mark_safe
 
-from corehq.apps.domain.models import Domain
 import corehq.apps.style.utils as style_utils
-from corehq.apps.hqwebapp.models import MaintenanceAlert
 from corehq.tabs import MENU_TABS
+
 
 register = template.Library()
 
@@ -129,43 +123,3 @@ def format_sidebar(context):
     return mark_safe(render_to_string(template, {
         'sections': sections
     }))
-
-
-@register.simple_tag
-def maintenance_alert():
-    try:
-        alert = (MaintenanceAlert.objects
-                 .filter(active=True)
-                 .order_by('-modified'))[0]
-    except IndexError:
-        return ''
-    else:
-        return format_html(
-            '<div class="alert alert-warning" style="text-align: center; margin-bottom: 0;">{}</div>',
-            mark_safe(alert.html),
-        )
-
-
-@register.filter
-def aliased_language_name(lang_code):
-    """
-    This is needed since we use non-standard language codes as alias, for e.g. 'fra' instead of 'fr' for French
-    """
-    try:
-        return language_name(lang_code)
-    except KeyError:
-        for code, name in settings.LANGUAGES:
-            if code == lang_code:
-                return name
-        raise KeyError('Unknown language code %s' % lang_code)
-
-
-@register.simple_tag(takes_context=True)
-def prelogin_url(context, urlname):
-    """
-    A prefix aware url tag replacement for prelogin URLs
-    """
-    if context.get('url_uses_prefix', False) and context.get('LANGUAGE_CODE', False):
-        return reverse(urlname, args=[context['LANGUAGE_CODE']])
-    else:
-        return reverse(urlname)
